Fix redirect for users trying to access pages without permissios

+++ events/views.py
@@ -1,7 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
-from django.contrib.auth.decorators import login_required
 
 from .models import Events
 from .forms import EventForm
@@ -30,7 +29,7 @@ def events_list_view(request):
             query = request.GET['q']
         if not query:
             messages.error(request, 'Please enter some search criteria!')
-            return redirect(reverse('events/list'))
+            return redirect(reverse('events_list_view'))
 
         queries = Q(name__icontains=query) | Q(description__icontains=query) | Q(location__icontains=query)
         events = events.filter(queries)
@@ -43,9 +42,11 @@ def events_list_view(request):
     return render(request, 'events/events_list_view.html', context)
 
 
-@login_required
 def manage_events(request):
     """Similar to event list view but with editing options"""
+    if not request.user.is_superuser:
+        messages.error(request, "You do not have the necessary permissions to access this page.")
+        return redirect('home')
 
     events = Events.objects.all().order_by('start_date')
 
@@ -58,8 +59,11 @@ def manage_events(request):
     return render(request, template, context)
 
 
-@login_required
 def add_event(request):
+    if not request.user.is_superuser:
+        messages.error(request, "You do not have the necessary permissions to perform this action.")
+        return redirect('home')
+    
     if request.method == 'POST':
         form = EventForm(request.POST, request.FILES)
         if form.is_valid():
@@ -75,9 +79,12 @@ def add_event(request):
     }
     return render(request, 'events/edit_event.html', context)
 
-# Update Event View
-@login_required
+
 def edit_event(request, event_id):
+    if not request.user.is_superuser:
+        messages.error(request, "You do not have the necessary permissions to perform this action.")
+        return redirect('home')
+    
     event = get_object_or_404(Events, id=event_id)
     if request.method == 'POST':
         form = EventForm(request.POST, request.FILES, instance=event)
@@ -96,9 +103,11 @@ def edit_event(request, event_id):
     return render(request, 'events/edit_event.html', context)
 
 
-# Delete Event View
-@login_required
 def delete_event(request, event_id):
+    if not request.user.is_superuser:
+        messages.error(request, "You do not have the necessary permissions to perform this action.")
+        return redirect('home')
+    
     event = get_object_or_404(Events, id=event_id)
     event.delete()
     messages.success(request, 'Event deleted successfully!')
